requirements.txt
pandas
google-generativeai
markdown
weasyprint
reportlab
fpdf
streamlit
